requirements.txt
torch
torchvision
Flask
numpy
Pillow
albumentations